last changes

+++ AuthService.Core/Interfaces/UserLoginResponse.cs
@@ -8,7 +8,7 @@ namespace AuthService.Core.Interfaces
 {
     public class UserLoginResponse
     {
-        public int UserID { get; set; }
+        public long UserID { get; set; }
         public string Username { get; set; } = string.Empty;
         public string RoleName { get; set; } = string.Empty;
         public string Token { get; set; } = string.Empty;
